move existing image utils to the single point where they are used
